refactor: move token-id to a separate crate
