+++ core/src/lib.rs
@@ -8,7 +8,6 @@ pub mod fees;
 pub mod intents;
 mod nonce;
 pub mod payload;
-pub mod token_id;
 
 pub use self::{deadline::*, error::*, nonce::*};
 
@@ -17,4 +16,5 @@ pub use defuse_erc191 as erc191;
 pub use defuse_nep413 as nep413;
 pub use defuse_sep53 as sep53;
 pub use defuse_tip191 as tip191;
+pub use defuse_token_id as token_id;
 pub use defuse_ton_connect as ton_connect;
