+++ token-id/src/nep141.rs
@@ -3,7 +3,7 @@ use std::{fmt, str::FromStr};
 use near_sdk::{AccountId, AccountIdRef, near};
 use serde_with::{DeserializeFromStr, SerializeDisplay};
 
-use crate::token_id::error::TokenIdError;
+use crate::error::TokenIdError;
 
 #[cfg(any(feature = "arbitrary", test))]
 use arbitrary_with::{Arbitrary, As};
