+++ token-id/src/nep171.rs
@@ -1,9 +1,11 @@
+pub use near_contract_standards::non_fungible_token::TokenId;
+
 use std::{fmt, str::FromStr};
 
 use near_sdk::{AccountId, AccountIdRef, near};
 use serde_with::{DeserializeFromStr, SerializeDisplay};
 
-use crate::token_id::{MAX_ALLOWED_TOKEN_ID_LEN, error::TokenIdError};
+use crate::{MAX_ALLOWED_TOKEN_ID_LEN, error::TokenIdError};
 
 #[cfg(any(feature = "arbitrary", test))]
 use arbitrary_with::{Arbitrary, As, LimitLen};
@@ -13,7 +15,7 @@ use defuse_near_utils::arbitrary::ArbitraryAccountId;
 #[cfg_attr(any(feature = "arbitrary", test), derive(Arbitrary))]
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, SerializeDisplay, DeserializeFromStr)]
 #[near(serializers = [borsh])]
-pub struct Nep245TokenId {
+pub struct Nep171TokenId {
     #[cfg_attr(
         any(feature = "arbitrary", test),
         arbitrary(with = As::<ArbitraryAccountId>::arbitrary),
@@ -24,21 +26,18 @@ pub struct Nep245TokenId {
         any(feature = "arbitrary", test),
         arbitrary(with = As::<LimitLen<MAX_ALLOWED_TOKEN_ID_LEN>>::arbitrary),
     )]
-    mt_token_id: defuse_nep245::TokenId,
+    nft_token_id: TokenId,
 }
 
-impl Nep245TokenId {
-    pub fn new(
-        contract_id: AccountId,
-        mt_token_id: defuse_nep245::TokenId,
-    ) -> Result<Self, TokenIdError> {
-        if mt_token_id.len() > MAX_ALLOWED_TOKEN_ID_LEN {
-            return Err(TokenIdError::TokenIdTooLarge(mt_token_id.len()));
+impl Nep171TokenId {
+    pub fn new(contract_id: AccountId, nft_token_id: TokenId) -> Result<Self, TokenIdError> {
+        if nft_token_id.len() > MAX_ALLOWED_TOKEN_ID_LEN {
+            return Err(TokenIdError::TokenIdTooLarge(nft_token_id.len()));
         }
 
         Ok(Self {
             contract_id,
-            mt_token_id,
+            nft_token_id,
         })
     }
 
@@ -47,30 +46,30 @@ impl Nep245TokenId {
         &self.contract_id
     }
 
-    pub const fn mt_token_id(&self) -> &defuse_nep245::TokenId {
-        &self.mt_token_id
+    pub const fn nft_token_id(&self) -> &TokenId {
+        &self.nft_token_id
     }
 
-    pub fn into_contract_id_and_mt_token_id(self) -> (AccountId, defuse_nep245::TokenId) {
-        (self.contract_id, self.mt_token_id)
+    pub fn into_contract_id_and_nft_token_id(self) -> (AccountId, TokenId) {
+        (self.contract_id, self.nft_token_id)
     }
 }
 
-impl std::fmt::Debug for Nep245TokenId {
+impl std::fmt::Debug for Nep171TokenId {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}:{}", self.contract_id(), self.mt_token_id())
+        write!(f, "{}:{}", self.contract_id(), self.nft_token_id())
     }
 }
 
-impl std::fmt::Display for Nep245TokenId {
+impl std::fmt::Display for Nep171TokenId {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Debug::fmt(&self, f)
     }
 }
 
-impl FromStr for Nep245TokenId {
+impl FromStr for Nep171TokenId {
     type Err = TokenIdError;
 
     fn from_str(data: &str) -> Result<Self, Self::Err> {
@@ -84,7 +83,6 @@ impl FromStr for Nep245TokenId {
 #[cfg(test)]
 mod tests {
     use super::*;
-
     use arbitrary::Unstructured;
     use arbitrary_with::UnstructuredExt;
     use defuse_test_utils::random::{make_arbitrary, random_bytes};
@@ -92,9 +90,9 @@ mod tests {
 
     #[rstest]
     #[trace]
-    fn display_from_str_roundtrip(#[from(make_arbitrary)] token_id: Nep245TokenId) {
+    fn display_from_str_roundtrip(#[from(make_arbitrary)] token_id: Nep171TokenId) {
         let s = token_id.to_string();
-        let got: Nep245TokenId = s.parse().unwrap();
+        let got: Nep171TokenId = s.parse().unwrap();
         assert_eq!(got, token_id);
     }
 
@@ -104,7 +102,7 @@ mod tests {
         let contract_id = u.arbitrary_as::<_, ArbitraryAccountId>().unwrap();
         let token_id: String = u.arbitrary().unwrap();
 
-        let r = Nep245TokenId::new(contract_id, token_id.clone());
+        let r = Nep171TokenId::new(contract_id, token_id.clone());
         if token_id.len() > MAX_ALLOWED_TOKEN_ID_LEN {
             assert!(matches!(r.unwrap_err(), TokenIdError::TokenIdTooLarge(_)));
         } else {
